Added greens function for blob

+++ src/main/java/com/stephenomalley/space/Point3D.java
@@ -5,30 +5,30 @@ import java.awt.geom.Point2D;
 /**
  * Created by kida_ on 14/10/2016.
  */
-public class Point3D extends Point2D.Float {
+public class Point3D extends Point2D.Double {
 
-    public float z;
+    public double z;
 
     public Point3D() {
     }
 
-    public Point3D(float x, float y, float z) {
+    public Point3D(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public void setLocation(float x, float y, float z) {
+    public void setLocation(double x, double y, double z) {
         this.x = x;
         this.y = y;
         this.z = z;
     }
 
-    public float getZ() {
+    public double getZ() {
         return z;
     }
 
-    public void setZ(float z) {
+    public void setZ(double z) {
         this.z = z;
     }
